+++ openBCIDebug.js
@@ -1,4 +1,5 @@
-var Buffer = require('safe-buffer').Buffer;
+const Buffer = require('buffer/').Buffer;
+
 module.exports = {
   debugBytes
 };